Make TraceUtil work with any TracingEngine.
TraceUtil previously was developed to work with JaegerTracingEngine.
With the introduction of NoopTracingEngine and LoggingTracingEngine this
is no longer possible.

Since every tracing engine was implementing three interfaces there was
no way to make TraceUtil#instance return a generic TracingEngine. As all engines implemented the same three interfaces
(TracingOpen, TracingOpenWithId, TracingOpenWithContext), a generic
TracingEngine interface implementing the aforementioned APIs was added to tracing-lib, and all engines were refactored to implement this new interface.

TraceUtil was then refactored to depend on the new TracingEngine API
instead.