Add sampling functionality to the tracing library
With sampling we can reduce the tracing overhead by limiting the
percentage of requests that are traced.

Since Jaeger's sampling functionality does not ignore the
instrumentation code we had to implement our own sampling.

This is done by generating a thread local random every time we call
addToTrace and if this value is below the sampling threshold a baggage
item ("sampled" : "True") is added to the span. From then on if the
active span's baggage does not contain that item, the instrumentation
code will be ignored and when necessary Noop versions of Span, Scope and
SpanContext are returned.